requirements.txt
httplib2
Django
djangorestframework
django-filter
gunicorn
celery
redis
jinja2
google-auth-oauthlib
google-api-python-client
dropbox
